+++ src/Symfony/Component/TypeInfo/Type/CollectionType.php
@@ -18,16 +18,16 @@ use Symfony\Component\TypeInfo\TypeIdentifier;
 /**
  * Represents a key/value collection type.
  *
- * It proxies every method to the main type and adds methods related to key and value types.
- *
  * @author Mathias Arlaud <mathias.arlaud@gmail.com>
  * @author Baptiste Leduc <baptiste.leduc@gmail.com>
  *
  * @template T of BuiltinType<TypeIdentifier::ARRAY>|BuiltinType<TypeIdentifier::ITERABLE>|ObjectType|GenericType
  *
+ * @implements WrappingTypeInterface<T>
+ *
  * @experimental
  */
-final class CollectionType extends Type
+final class CollectionType extends Type implements WrappingTypeInterface
 {
     /**
      * @param T $type
@@ -36,7 +36,7 @@ final class CollectionType extends Type
         private readonly BuiltinType|ObjectType|GenericType $type,
         private readonly bool $isList = false,
     ) {
-        if ($this->isList()) {
+        if ($this->isList) {
             $keyType = $this->getCollectionKeyType();
 
             if (!$keyType instanceof BuiltinType || TypeIdentifier::INT !== $keyType->getTypeIdentifier()) {
@@ -45,32 +45,14 @@ final class CollectionType extends Type
         }
     }
 
-    public function getBaseType(): BuiltinType|ObjectType
-    {
-        return $this->getType()->getBaseType();
-    }
-
-    /**
-     * @return T
-     */
-    public function getType(): BuiltinType|ObjectType|GenericType
+    public function getWrappedType(): Type
     {
         return $this->type;
-    }
-
-    public function isA(TypeIdentifier|string $subject): bool
-    {
-        return $this->getType()->isA($subject);
     }
 
     public function isList(): bool
     {
         return $this->isList;
-    }
-
-    public function asNonNullable(): self
-    {
-        return $this;
     }
 
     public function getCollectionKeyType(): Type
@@ -106,15 +88,5 @@ final class CollectionType extends Type
     public function __toString(): string
     {
         return (string) $this->type;
-    }
-
-    /**
-     * Proxies all method calls to the original type.
-     *
-     * @param list<mixed> $arguments
-     */
-    public function __call(string $method, array $arguments): mixed
-    {
-        return $this->type->{$method}(...$arguments);
     }
 }