+++ src/Symfony/Component/TypeInfo/Type/TemplateType.php
@@ -11,9 +11,7 @@
 
 namespace Symfony\Component\TypeInfo\Type;
 
-use Symfony\Component\TypeInfo\Exception\LogicException;
 use Symfony\Component\TypeInfo\Type;
-use Symfony\Component\TypeInfo\TypeIdentifier;
 
 /**
  * Represents a template placeholder, such as "T" in "Collection<T>".
@@ -27,18 +25,8 @@ final class TemplateType extends Type
 {
     public function __construct(
         private readonly string $name,
-        private readonly Type $bound,
+        private readonly Type $boundType,
     ) {
-    }
-
-    public function getBaseType(): BuiltinType|ObjectType
-    {
-        throw new LogicException(\sprintf('Cannot get base type on "%s" template type.', $this));
-    }
-
-    public function isA(TypeIdentifier|string $subject): bool
-    {
-        return false;
     }
 
     public function getName(): string
@@ -46,14 +34,9 @@ final class TemplateType extends Type
         return $this->name;
     }
 
-    public function getBound(): Type
+    public function getBoundType(): Type
     {
-        return $this->bound;
-    }
-
-    public function asNonNullable(): self
-    {
-        return $this;
+        return $this->boundType;
     }
 
     public function __toString(): string