+++ src/Symfony/Component/TypeInfo/CHANGELOG.md
@@ -1,6 +1,15 @@
 CHANGELOG
 =========
 
+7.2
+---
+
+ * Add `Type::traverse` and `Type::isScalar` methods
+ * Add `WrappingTypeInterface` and `CompositeTypeInterface` type interfaces
+ * Add `NullableType` type class
+ * Remove `Type::asNonNullable`, `Type::getBaseType`, `Type::is`, and `Type::__call`, methods
+ * Remove `CompositeTypeTrait::atLeastOneTypeIs` and `CompositeTypeTrait::everyTypeIs` methods
+
 7.1
 ---